src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/parent-tree-scope-in-shadow-expected.txt

CONSOLE ERROR: line 14: Uncaught TypeError: div.createShadowRoot is not a function
Creating shadow dom subtrees from top to bottom.
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
